Create SQL Create Stored Procedure Drill
Drill was to design a stored procedure that contains two  joins and two parameters...using the database provided called AdventureWorks.  As a manager, I frequently had to monitor the vacation usage of my employees, so I decided to create a stored procedure that could be run  to find out who had how much vacation time left, depending on the JobTitle I entered and the # of vacation hours left I was concerned about.